django/requirements.txt
Django==5.1.7
djangorestframework==3.15.2  # for api
django-filter==25.1  # for filtering with djangorestframework
django-import-export==4.3.7  # for bulk import, export
django-mptt==0.17.0  # tree structure model for folder models
django-redis==5.4.0  # for server-side caching with redis
daphne==4.1.2  # asgi server
channels==4.2.0  # for asgi(websocket)
channels_redis==4.2.1  # for channels with redis
redis==5.2.1  # for django-redis and channels_redis
celery==5.4.0  # async process
pillow==11.1.0  # image processing
openslide-bin==4.0.0.6  # openslide binary
openslide-python==1.4.1  # openslide python interface
psycopg==3.2.6  # for postgresql
python-dotenv==1.1.0  # loading .env file
